finish refactor for map block. relates to #504
