+++ src/blocks/map/index.js
@@ -2,8 +2,10 @@
  * Internal dependencies
  */
 import './styles/editor.scss';
-import edit from './components/edit';
-import save from './components/save';
+import edit from './edit';
+import save from './save';
+import deprecated from './deprecated';
+import transforms from './transforms';
 import icons from './../../utils/icons';
 import metadata from './block.json';
 
@@ -11,146 +13,37 @@ import metadata from './block.json';
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
 
 /**
  * Block constants
  */
-const { name } = metadata;
+const { attributes, name } = metadata;
 
 const icon = icons.googleMap;
 
-const blockAttributes = {
-	address: {
-		type: 'string',
-	},
-	lat: {
-		type: 'string',
-	},
-	lng: {
-		type: 'string',
-	},
-	hasApiKey: {
-		type: 'boolean',
-		default: false,
-	},
-	pinned: {
-		type: 'boolean',
-		default: false,
-	},
-	height: {
-		type: 'number',
-		default: 400,
-	},
-	skin: {
-		type: 'string',
-		default: 'standard',
-	},
-	zoom: {
-		type: 'number',
-		default: 12,
-	},
-	iconSize: {
-		type: 'number',
-		default: 36,
-	},
-	mapTypeControl: {
-		type: 'boolean',
-		default: true,
-	},
-	zoomControl: {
-		type: 'boolean',
-		default: true,
-	},
-	streetViewControl: {
-		type: 'boolean',
-		default: true,
-	},
-	fullscreenControl: {
-		type: 'boolean',
-		default: true,
-	},
-	controls: {
-		type: 'boolean',
-		default: true,
-	},
-	hasError: {
-		type: 'string',
-	},
-};
+const title = __( 'Map' );
 
 const settings = {
-	title: __( 'Map' ),
+	title,
+
 	description: __( 'Add an address and drop a pin on a Google map.' ),
+
 	keywords: [ __( 'address' ), __( 'maps' ), __( 'google' ) ],
+
 	supports: {
 		align: [ 'wide', 'full' ],
 		coBlocksSpacing: true,
 	},
-	attributes: metadata.attributes,
-	transforms: {
-		from: [
-			{
-				type: 'prefix',
-				prefix: ':map',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-		],
-	},
+
+	attributes,
+
+	transforms,
+
 	edit,
+
 	save,
-	deprecated: [
-		{
-			attributes: blockAttributes,
-			save( { attributes } ) {
-				const {
-					address,
-					height,
-					lat,
-					lng,
-					skin,
-					zoom,
-					iconSize,
-					mapTypeControl,
-					zoomControl,
-					streetViewControl,
-					fullscreenControl,
-				} = attributes;
 
-				const backgroundStyles = {
-					minHeight: height ? height + 'px' : undefined,
-				};
-
-				const mapAttributes = {
-					address,
-					lat,
-					lng,
-					skin,
-					zoom,
-					iconSize,
-					mapTypeControl,
-					zoomControl,
-					streetViewControl,
-					fullscreenControl,
-				};
-
-				const attr = Object
-					.keys( mapAttributes )
-					.map( key => `/q${ key }/q:/q${ mapAttributes[ key ] }/q` )
-					.join( '||' );
-
-				const dataMap = { 'data-map-attr': attr };
-
-				return (
-					<div style={ backgroundStyles } { ...dataMap } />
-				);
-			},
-		},
-	],
+	deprecated,
 };
 
-export { name, icon, settings };
+export { name, title, icon, settings };
