+++ src/blocks/map/inspector.js
